+++ packages/gamut/src/Tag/index.tsx
@@ -1,130 +1,10 @@
 import { MiniDeleteIcon } from '@codecademy/gamut-icons';
-import {
-  Background,
-  css,
-  system,
-  theme,
-  useCurrentMode,
-  variant,
-} from '@codecademy/gamut-styles';
-import { StyleProps, variance } from '@codecademy/variance';
-import styled from '@emotion/styled';
+import { useCurrentMode } from '@codecademy/gamut-styles';
 import React from 'react';
 
-import { Box } from '../Box';
-import { ButtonProps } from '../Button';
-import { ButtonBase } from '../ButtonBase';
-import { Selectors } from '../ButtonBase/ButtonBase';
 import { Text } from '../Typography';
-
-const Outline = styled(Box)(
-  css({
-    // this is a bit of a hack as we don't have access to focus-visible from this component.  if you are not properly dismissing your tags you may see this primary colored outline after clicking X, but otherwise you should never hit this behavior.
-
-    borderRadius: '4px',
-    width: 'fit-content',
-    maxWidth: '100%',
-    '&:focus-within': {
-      outline: `2px solid ${theme.colors.primary}`,
-      outlineOffset: '2px',
-    },
-    '&:active': {
-      outlineColor: `transparent`,
-    },
-  })
-);
-
-const colorVariants = variant({
-  defaultVariant: 'default',
-  base: {
-    alignItems: 'center',
-    borderRadius: '4px',
-    display: 'flex',
-    height: '24px',
-    justifyContent: 'center',
-    width: 'fit-content',
-    maxWidth: '100%',
-  },
-  variants: {
-    default: {
-      bg: `background-current`,
-      color: 'text',
-    },
-    grey: {
-      bg: `navy-500`,
-      color: 'white',
-    },
-  },
-});
-
-const tagProps = variance.compose(
-  system.space,
-  system.layout,
-  system.typography
-);
-
-export interface BaseTagProps
-  extends StyleProps<typeof tagProps>,
-    StyleProps<typeof colorVariants> {}
-export interface ReadOnlyTagProps extends BaseTagProps {
-  /**
-   * If the DismissButton should be shown.
-   */
-  readonly: true;
-  /**
-   * ClickHandler for the DismissButton.
-   */
-  onDismiss?: never;
-}
-export interface DismissableTagProps extends BaseTagProps {
-  /**
-   * If the DismissButton should be shown.
-   */
-  readonly?: false;
-  /**
-   * ClickHandler for the DismissButton.
-   */
-  onDismiss: ButtonProps['onClick'];
-}
-
-export type TagProps = ReadOnlyTagProps | DismissableTagProps;
-
-const DeleteButton = styled(ButtonBase)(
-  variant({
-    defaultVariant: 'default',
-    base: {
-      alignItems: 'center',
-      border: 1,
-      borderColor: 'transparent',
-      borderRadiusRight: '4px',
-      color: 'currentColor',
-      display: 'flex',
-      height: '100%',
-      justifyContent: 'center',
-      minWidth: '24px',
-    },
-    variants: {
-      default: {
-        [Selectors.HOVER]: {
-          bg: 'background-hover',
-        },
-        [Selectors.FOCUS]: {
-          bg: 'background-selected',
-        },
-      },
-      grey: {
-        [Selectors.HOVER]: {
-          bg: 'navy-600',
-        },
-        [Selectors.FOCUS]: {
-          bg: 'navy-700',
-        },
-      },
-    },
-  })
-);
-
-const TagWrapper = styled(Background)<BaseTagProps>(tagProps, colorVariants);
+import { DeleteButton, Outline, TagWrapper } from './elements';
+import { TagProps } from './types';
 
 export const Tag: React.FC<TagProps> = ({
   children,
