+++ packages/gamut/src/Tag/types.tsx
@@ -1,0 +1,30 @@
+import { StyleProps } from '@codecademy/variance';
+
+import { ButtonProps } from '../Button';
+import { colorVariants, tagProps } from './elements';
+
+export interface BaseTagProps
+  extends StyleProps<typeof tagProps>,
+    StyleProps<typeof colorVariants> {}
+export interface ReadOnlyTagProps extends BaseTagProps {
+  /**
+   * If the DismissButton should be shown.
+   */
+  readonly: true;
+  /**
+   * ClickHandler for the DismissButton.
+   */
+  onDismiss?: never;
+}
+export interface DismissableTagProps extends BaseTagProps {
+  /**
+   * If the DismissButton should be shown.
+   */
+  readonly?: false;
+  /**
+   * ClickHandler for the DismissButton.
+   */
+  onDismiss: ButtonProps['onClick'];
+}
+
+export type TagProps = ReadOnlyTagProps | DismissableTagProps;
